fix : Java 17 > 11 문법으로 변경
