+++ build.gradle
@@ -5,6 +5,9 @@ plugins {
 group = 'camp.nextstep.edu'
 version = '1.0-SNAPSHOT'
 
+sourceCompatibility = 11
+targetCompatibility = 11
+
 repositories {
     mavenCentral()
 }
