--- src/test/java/persistence/sql/ddl/generator/example/PersonV2WithGeneratedValue.java
@@ -1,22 +1,0 @@
-package persistence.sql.ddl.generator.example;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class PersonV2WithGeneratedValue {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private Integer age;
-
-    private String email;
-
-}
