+++ src/test/java/persistence/sql/ddl/generator/fixture/PersonV1.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.generator.example;
+package persistence.sql.ddl.generator.fixture;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
