+++ src/test/java/persistence/sql/ddl/generator/fixture/PersonV2WithColumnName.java
@@ -1,16 +1,13 @@
-package persistence.sql.ddl.generator.example;
+package persistence.sql.ddl.generator.fixture;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
 @Entity
-public class PersonV2 {
+public class PersonV2WithColumnName {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(name = "nick_name")
@@ -19,7 +16,7 @@ public class PersonV2 {
     @Column(name = "old")
     private Integer age;
 
-    @Column(nullable = false)
+    @Column(name = "email_contact")
     private String email;
 
 }
