+++ src/test/java/persistence/sql/ddl/generator/fixture/PersonV3.java
@@ -1,15 +1,19 @@
-package persistence.sql.ddl.generator.example;
+package persistence.sql.ddl.generator.fixture;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
 
+@Table(name = "users")
 @Entity
-public class PersonV2WithColumnName {
+public class PersonV3 {
 
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(name = "nick_name")
@@ -18,7 +22,10 @@ public class PersonV2WithColumnName {
     @Column(name = "old")
     private Integer age;
 
-    @Column(name = "email_contact")
+    @Column(nullable = false)
     private String email;
+
+    @Transient
+    private Integer index;
 
 }
