+++ src/test/java/persistence/sql/ddl/generator/fixture/PersonV3WithTable.java
@@ -1,12 +1,12 @@
-package persistence.sql.ddl.generator.example;
+package persistence.sql.ddl.generator.fixture;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
 
+@Table(name = "users")
 @Entity
-public class PersonV3WithTransient {
+public class PersonV3WithTable {
 
     @Id
     private Long id;
@@ -16,8 +16,4 @@ public class PersonV3WithTransient {
     private Integer age;
 
     private String email;
-
-    @Transient
-    private Integer index;
-
 }
